pyproject.toml
[tool.poetry]
name = "portfolio"
version = "0.1.0"
description = ""
authors = ["nickmflorin <nickmflorin@gmail.com>"]
repository = "https://github.com/nickmflorin/portfolio-drf"
homepage = "https://github.com/nickmflorin/portfolio-drf"
packages = [
  { include = "portfolio", from = "src" },
]
license = "UNLICENSED"
include = [
  "README.rst",
  "pyproject.toml"
]

[tool.poetry.extras]
docs = [
  "sphinx",
  "sphinx-rtd-theme",
  "sphinxcontrib.documentedlist",
  "sphinxcontrib.httpdomain",
  "sphinxcontrib.httpexample"
]
debug = [ "ipython", "ipdb" ]

[tool.poetry.dependencies]
python = "^3.7"
pytest-django = "^3.8"
dj-database-url = "^0.5.0"
tox = "^3.14"
django-cors-headers = "^3.2"
pillow = "^7.0"
djangocms-admin-style = "^1.5"
djangorestframework = "^3.11"
django = "^3.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
